Conditionally run system upgrade (-Syu) before installing packages
On a fresh install it might happen that after priority files, including
the pacman config, are installed the pacman database is not up to date.

  : Applying configuration...
  :: Installing priority files...
  ::: Installing /etc/pacman.conf...
  :::: Done.
  ::: Done.
  :: Configuring packages...
  ::: Unpinning 4 unknown packages.
  warning: database file for 'community' does not exist (use '-Sy' to download)
  warning: database file for 'multilib' does not exist (use '-Sy' to download)

This can later result in issues with finding the correct package during
the "Apply packages" step.

  error: target not found: lib32-vulkan-intel
  error: target not found: lib32-vulkan-mesa-layers
  error: target not found: steam
  error: target not found: wine

To avoid this issue ask the user if he wishes to do as pacman suggests
and update the database before "Apply packages". To avoid an
inconsistent state upgrade packages for which database entries were
updated, in other words perform a system upgrade.